refactor(animation): refactor Animation class to a separate module
- Moved the `Animation` class from `rive.ts` to a new file `Animation.ts` under `js/src/animation/`.
- Added a new `index.ts` file in the `js/src/animation/` directory to export the `Animation` class.
- Updated imports in `rive.ts` to use the newly created `Animation` module.

This refactor improves code organization by separating the `Animation` class into its own module, making the codebase more modular and easier to maintain.
