+++ js/src/animation/index.ts
@@ -1,0 +1,1 @@
+export { Animation } from "./Animation";
