+++ js/src/rive.ts
@@ -1,5 +1,6 @@
 import * as rc from "./rive_advanced.mjs";
 import * as packageData from "package.json";
+import { Animation } from './animation';
 import { registerTouchInteractions, sanitizeUrl, BLANK_URL } from "./utils";
 
 // Note: Re-exporting a few types from rive_advanced.mjs to expose for high-level
@@ -259,92 +260,6 @@ export class RuntimeLoader {
 
 // #endregion
 
-// #region animations
-
-// Wraps animations and instances from the runtime and keeps track of playback
-// state
-class Animation {
-  public loopCount = 0;
-  public readonly instance: rc.LinearAnimationInstance;
-
-  // Time to which the animation should move to on the next render
-  public scrubTo: number | null = null;
-
-  /**
-   * Constructs a new animation
-   * @constructor
-   * @param {any} animation: runtime animation object
-   * @param {any} instance: runtime animation instance object
-   */
-  constructor(
-    private animation: rc.LinearAnimation,
-    private artboard: rc.Artboard,
-    runtime: rc.RiveCanvas,
-    public playing: boolean,
-  ) {
-    this.instance = new runtime.LinearAnimationInstance(animation, artboard);
-  }
-
-  // Returns the animation's name
-  public get name(): string {
-    return this.animation.name;
-  }
-
-  // Returns the animation's current time
-  public get time(): number {
-    return this.instance.time;
-  }
-
-  // Sets the animation's current time
-  public set time(value: number) {
-    this.instance.time = value;
-  }
-
-  // Returns the animation's loop type
-  public get loopValue(): number {
-    return this.animation.loopValue;
-  }
-
-  /**
-   * Advances the animation by the give time. If the animation needs scrubbing,
-   * time is ignored and the stored scrub value is used.
-   * @param time the time to advance the animation by if no scrubbing required
-   */
-  public advance(time: number) {
-    if (this.scrubTo === null) {
-      this.instance.advance(time);
-    } else {
-      this.instance.time = 0;
-      this.instance.advance(this.scrubTo);
-      this.scrubTo = null;
-    }
-  }
-
-  /**
-   * Apply interpolated keyframe values to the artboard. This should be called after calling
-   * .advance() on an animation instance so that new values are applied to properties.
-   *
-   * Note: This does not advance the artboard, which updates all objects on the artboard
-   * @param mix - Mix value for the animation from 0 to 1
-   */
-  public apply(mix: number) {
-    this.instance.apply(mix);
-  }
-
-  public get needsScrub(): boolean {
-    return this.scrubTo !== null;
-  }
-
-  /**
-   * Deletes the backing Wasm animation instance; once this is called, this
-   * animation is no more.
-   */
-  public cleanup() {
-    this.instance.delete();
-  }
-}
-
-// #endregion
 
 // #region state machines
